refactor: Fixed UI misbehaviour
